+++ src/lib/services/evm/index.ts
@@ -11,7 +11,6 @@ import { isHexWalletAddress } from "lib/utils";
 import {
   getEvmCodesByAddress,
   getEvmContractInfoSequencer,
-  getEvmDenomByAddressLcd,
   getEvmParams,
 } from "./lcd";
 
@@ -58,20 +57,6 @@ export const useEvmContractInfoSequencer = (address: HexAddr20) => {
       address,
     ],
     async () => getEvmContractInfoSequencer(lcdEndpoint, prefix, address),
-    {
-      retry: 1,
-      refetchOnWindowFocus: false,
-      enabled: address && isHexWalletAddress(address),
-    }
-  );
-};
-
-export const useEvmDenomByAddressLcd = (address: HexAddr20) => {
-  const lcdEndpoint = useLcdEndpoint();
-
-  return useQuery(
-    [CELATONE_QUERY_KEYS.EVM_DENOM_BY_ADDRESS_LCD, lcdEndpoint, address],
-    async () => getEvmDenomByAddressLcd(lcdEndpoint, address),
     {
       retry: 1,
       refetchOnWindowFocus: false,